requirements.txt
# 核心依赖
matplotlib>=3.5.0
numpy>=1.20.0

# 配置文件支持
PyYAML>=6.0

# 系统信息获取
psutil>=5.8.0

# 开发和测试依赖（可选）
pytest>=7.0.0
pytest-cov>=4.0.0